Separate Orientation and Misorientation tests
